+++ package/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.8
+
+- Support to set max character on TEXT input field
+
+## 1.1.7
+
+- Fixing missing required paremeter in Map input field
+
 ## 1.1.6
 
 - Bugfix: Coordinate input field